Rename internal dir
