+++ otel/internal/utils/mapotelstatus.go
