+++ otel/internal/utils/spanattributes.go
